+++ src/app/app.module.ts
@@ -1,31 +1,26 @@
 import { NgModule } from "@angular/core";
 import { AngularFireModule } from "@angular/fire";
 import { AngularFirestoreModule } from "@angular/fire/firestore";
-import { FlexLayoutModule } from "@angular/flex-layout";
 import { BrowserModule } from "@angular/platform-browser";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 
 import { AppComponent } from "./app.component";
-import { CounterComponent } from "./counter/counter.component";
-import { CounterDialogComponent } from "./counter-dialog/counter-dialog.component";
-import { CountersService } from "./service/counters.service";
 import { MaterialModule } from "./material.module";
 import { environment } from "src/environments/environment";
-import { FormsModule } from "@angular/forms";
+import { CounterDashboardModule } from "./counter-dashboard/counter-dashboard.module";
+import { ToolbarComponent } from "./toolbar/toolbar.component";
 
 @NgModule({
-  declarations: [AppComponent, CounterComponent, CounterDialogComponent],
+  declarations: [AppComponent, ToolbarComponent],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
     MaterialModule,
     AngularFireModule.initializeApp(environment.firebase),
     AngularFirestoreModule,
-    FlexLayoutModule,
-    FormsModule
+    CounterDashboardModule
   ],
-  providers: [CountersService],
-  bootstrap: [AppComponent],
-  entryComponents: [CounterDialogComponent]
+  providers: [],
+  bootstrap: [AppComponent]
 })
 export class AppModule {}