+++ src/app/counter-dashboard/counter-dashboard.component.spec.ts
@@ -1,20 +1,20 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { CounterComponent } from './counter.component';
+import { CounterDashboardComponent } from './counter-dashboard.component';
 
-describe('CounterComponent', () => {
-  let component: CounterComponent;
-  let fixture: ComponentFixture<CounterComponent>;
+describe('CounterDashboardComponent', () => {
+  let component: CounterDashboardComponent;
+  let fixture: ComponentFixture<CounterDashboardComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CounterComponent ]
+      declarations: [ CounterDashboardComponent ]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(CounterComponent);
+    fixture = TestBed.createComponent(CounterDashboardComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });