+++ src/app/material.module.ts
@@ -5,7 +5,7 @@ import {
   MatFormFieldModule,
   MatIconModule,
   MatInputModule,
-  MatToolbarModule,
+  MatToolbarModule
 } from "@angular/material";
 import { NgModule } from "@angular/core";
 
@@ -17,7 +17,7 @@ import { NgModule } from "@angular/core";
     MatFormFieldModule,
     MatIconModule,
     MatInputModule,
-    MatToolbarModule,
+    MatToolbarModule
   ],
   exports: [
     MatButtonModule,
@@ -26,7 +26,7 @@ import { NgModule } from "@angular/core";
     MatFormFieldModule,
     MatIconModule,
     MatInputModule,
-    MatToolbarModule,
+    MatToolbarModule
   ]
 })
 export class MaterialModule {}